requirements.txt
numpy
matplotlib
sympy
scipy
pandas
pillow
opencv-python
pytesseract
pytest
rich
